# MainPostsAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MainPostsAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Post/get_posts": {
      "get": {
        "tags": [
          "Post"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PostVO"
              }
            }
          }
        }
      }
    },
    "/Post/get_projects": {
      "get": {
        "tags": [
          "Post"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PostVO"
              }
            }
          }
        }
      }
    },
    "/Post/get_post/{id}": {
      "get": {
        "tags": [
          "Post"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PostVO"
            }
          }
        }
      }
    },
    "/Post/get_post_itens/{id}": {
      "get": {
        "tags": [
          "Post"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PostItemVO"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ItemType": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4
      ],
      "type": "integer"
    },
    "PostItem": {
      "required": [
        "postType"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "postId": {
          "format": "int64",
          "type": "integer"
        },
        "postType": {
          "$ref": "#/definitions/ItemType"
        },
        "subtitle": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "adress": {
          "type": "string"
        }
      }
    },
    "PostItemVO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "postId": {
          "format": "int64",
          "type": "integer"
        },
        "postType": {
          "$ref": "#/definitions/ItemType"
        },
        "subtitle": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "adress": {
          "type": "string"
        }
      }
    },
    "PostVO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "postTitle": {
          "type": "string"
        },
        "postImage": {
          "type": "string"
        },
        "postDescription": {
          "type": "string"
        },
        "isAProject": {
          "type": "boolean"
        },
        "postActive": {
          "type": "boolean"
        },
        "postDate": {
          "format": "date-time",
          "type": "string"
        },
        "postItens": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PostItem"
          }
        }
      }
    }
  }
}